+++ user_management/models/mixins.py
@@ -1,3 +1,4 @@
+import pytz
 from django.contrib.auth.models import BaseUserManager
 from django.contrib.auth.tokens import default_token_generator
 from django.contrib.sites.models import Site
@@ -9,6 +10,7 @@ from django.utils.http import urlsafe_base64_encode
 from django.utils.translation import ugettext_lazy as _
 
 from user_management.utils import notifications
+from .utils import timezone_choices
 
 
 class UserManager(BaseUserManager):
@@ -49,6 +51,16 @@ class DateJoinedUserMixin(models.Model):
         verbose_name=_('date joined'),
         default=timezone.now,
         editable=False,
+    )
+
+    class Meta:
+        abstract = True
+
+
+class TimeZoneMixin(models.Model):
+    timezone = models.CharField(
+        max_length=255,
+        choices=timezone_choices(pytz.common_timezones),
     )
 
     class Meta: