+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+* Add `TimeZoneMixin` for custom `User` models.
+
 ## v14.0.0
 
 * Clarify error message when your old and new passwords match, you will need to update translations.